requirements.txt
psycopg2-binary
pandas
gspread
oauth2client
